Minor code refactor
