+++ src/main/java/ogc/rs/processes/ProcessesRunnerImpl.java
@@ -15,7 +15,7 @@ import io.vertx.sqlclient.Tuple;
 import ogc.rs.common.DataFromS3;
 import ogc.rs.processes.collectionAppending.CollectionAppendingProcess;
 import ogc.rs.processes.collectionOnboarding.CollectionOnboardingProcess;
-import ogc.rs.processes.preSignedURLGeneration.PreSignedURLGenerationProcess;
+import ogc.rs.processes.s3PreSignedURLGeneration.S3PreSignedURLGenerationProcess;
 import ogc.rs.processes.util.Status;
 import ogc.rs.processes.util.UtilClass;
 import org.apache.logging.log4j.LogManager;
@@ -85,8 +85,8 @@ public class ProcessesRunnerImpl implements ProcessesRunnerService {
           case "CollectionAppending":
             processService = new CollectionAppendingProcess(pgPool, webClient, config,getS3Object(config),vertx);
             break;
-          case "PreSignedURLGeneration":
-            processService = new PreSignedURLGenerationProcess(pgPool, config);
+          case "S3PreSignedURLGeneration":
+            processService = new S3PreSignedURLGenerationProcess(pgPool, config);
         }
 
         ProcessService finalProcessService = processService;
